requirements.txt
https://www.djangoproject.com/download/1.7c2/tarball/
psycopg2
django-websocket-redis
ago
pytz